Since we are doing almost ruby on rails then why not application helpers. application helper selected methods required and exported in courseInstance helper and exported to the contloller
