+++ backend/server/controllers/index.js
@@ -1,3 +1,5 @@
+
+
 const courses = require('./courses')
 const courseInstances = require('./courseinstances')
 const studentInstances = require('./studentinstances')
