+++ backend/server/routes/courseInstanceRouter.js
@@ -1,27 +1,7 @@
 const courseInstanceController = require('../controllers').courseInstances
-//const jwt = require('jsonwebtoken')
-/*let express = require('express')*/
 
 module.exports = (app) => {
-  /*const apiRoutes = express.Router()
-  apiRoutes.use(function (req, res, next) {
-    const token = req.body.token || req.query.token
-    if (token) {
-      jwt.verify(token, app.get(process.env.SECRET), function (err, decoded) {
-        if (err) {
-          return res.json({ success: false, message: 'Failed to authenticate token' })
-        } else {
-          req.decoded = decoded
-          next()
-        }
-      })
-    } else {
-      return res.status(403).send({
-        success: false,
-        message: 'No token provided'
-      })
-    }
-  })*/
+
   app.post('/api/courseinstances/update', courseInstanceController.getNew) //For updating DB with data from opetushallitus
   app.post('/api/courseinstances/update/next', courseInstanceController.getNewer)
   app.post('/api/course/register/:ohid', courseInstanceController.testi)
